(feat) Added separate LIC tests
